+++ src/Credfeto.Cache.Proxy.Server/Extensions/UriExtensions.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Credfeto.Cache.Proxy.Server.Extensions;
 
-internal static class UriExtensions
+public static class UriExtensions
 {
     public static string CleanUri(this Uri uri)
     {