+++ src/Credfeto.Cache.Proxy.Server/Extensions/EnumExtensions.cs
@@ -4,4 +4,4 @@ using Credfeto.Enumeration.Source.Generation.Attributes;
 namespace Credfeto.Cache.Proxy.Server.Extensions;
 
 [EnumText(typeof(HttpStatusCode))]
-internal static partial class EnumExtensions;
+public static partial class EnumExtensions;