+++ src/Credfeto.Cache.Proxy.Server/Helpers/Endpoints.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Builder;
 
 namespace Credfeto.Cache.Proxy.Server.Helpers;
 
-internal static partial class Endpoints
+public static partial class Endpoints
 {
     public static WebApplication ConfigureEndpoints(this WebApplication app)
     {