+++ src/Credfeto.Cache.Proxy.Server/Helpers/StartupBanner.cs
@@ -5,7 +5,7 @@ namespace Credfeto.Cache.Proxy.Server.Helpers;
 
 // https://www.figlet.org/examples.html
 [GenerateFiggleText("Banner", "basic", "Cache Proxy")]
-internal static partial class StartupBanner
+public static partial class StartupBanner
 {
     public static void Show()
     {