+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated FunFair.CodeAnalysis to 7.2.1.2035
 - Dependencies - Updated Meziantou.Analyzer to 3.0.102
 - Dependencies - Updated Microsoft.Extensions to 10.0.9
+- Dependencies - Updated Credfeto.Docker.HealthCheck.Http.Client to 0.0.67.857
 ### Deprecated
 ### Removed
 ### Deployment Changes