+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated Credfeto.Enumeration to 1.2.146.1978
+- Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.129.1310
 ### Deprecated
 ### Removed
 ### Deployment Changes